# Create startup snapshots at build time