--- a/crates/funcgg-runtime/src/lib.rs
+++ b/crates/funcgg-runtime/src/lib.rs
@@ -3,4 +3,5 @@ mod loader;
 mod sandbox;
 
 pub mod http;
+pub mod snapshot;
 pub use sandbox::Sandbox;