--- a/crates/funcgg-worker/Cargo.toml
+++ b/crates/funcgg-worker/Cargo.toml
@@ -23,3 +23,6 @@ tower-http = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 uuid = { workspace = true }
+
+[build-dependencies]
+funcgg-runtime = { path = "../funcgg-runtime" }