--- a/crates/funcgg-worker/src/workers/worker.rs
+++ b/crates/funcgg-worker/src/workers/worker.rs
@@ -6,6 +6,9 @@ use uuid::Uuid;
 use super::pool::StateChange;
 use funcgg_runtime::{Sandbox, http};
 
+static STARTUP_SNAPSHOT: &[u8] =
+    include_bytes!(concat!(env!("OUT_DIR"), "/FUNCGG_RUNTIME_SNAPSHOT.bin"));
+
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct WorkerRequest {
     pub id: Uuid,
@@ -56,7 +59,7 @@ impl Worker {
     }
 
     async fn process_request(&self, request: WorkerRequest) -> Result<http::Response, String> {
-        let mut sandbox = match Sandbox::new(request.id) {
+        let mut sandbox = match Sandbox::new(request.id, Some(STARTUP_SNAPSHOT)) {
             Ok(rt) => rt,
             Err(e) => {
                 tracing::error!(