This gives a nice start boost. I took a sample of "hello world" requests to compare timings from request received to isolate initialized. These timings do not include the actual runtime of the user code.

Before startup snapshot:

```
min: 44.461708ms, max: 75.167375ms, avg: 52.93726154385965ms
```

After snapshotting:

```
min: 7.710125ms, max: 23.561291ms, avg: 11.90164068888889ms
```